--- src/components/button/custom-button/CustomButton.jsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import './custom-button.scss';
-
-const CustomButton = ({
-  children,
-  isGoogleSignIn,
-  inverted,
-  ...otherProps
-}) => (
-  <button
-    className={`${inverted ? 'inverted' : ''} ${
-      isGoogleSignIn ? 'google-sign-in' : ''
-    } custom-button`}
-    {...otherProps}
-  >
-    {children}
-  </button>
-);
-
-export default CustomButton;
