+++ src/components/directory/directory.scss
