+++ src/features/homepage/HomePage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Directory from '../../features/directory/Directory';
+import Directory from '../../components/directory/Directory';
 import './homepage.scss';
 
 const HomePage = () => {
